# public/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Starter",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1"
    }
  ],
  "security": [
    {
      "http": []
    }
  ],
  "paths": {
...
    "/webhook-endpoints/{endpoint}": {
      "get": {
        "operationId": "v1.webhook-endpoints.show",
        "summary": "Retrieve a webhook endpoint",
        "tags": ["Webhook endpoints", "WebhookEndpoint"],
        "parameters": [
          {
            "name": "endpoint",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "`WebhookEndpointResource`",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/WebhookEndpointResource"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationException"
          }
        }
      },
      "put": {
        "operationId": "v1.webhook-endpoints.update",
        "summary": "Update a webhook endpoint",
        "tags": ["Webhook endpoints", "WebhookEndpoint"],
        "parameters": [
          {
            "name": "endpoint",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "description": "https only, and pointed outside our own network: this is the one URL a customer chooses\nthat our servers then request, which makes it an SSRF primitive without the second check.\nA delivery is signed but not encrypted, and the payload describes the customer's own data.",
                    "maxLength": 2000
                  },
                  "description": {
                    "type": ["string", "null"],
                    "maxLength": 255
                  },
                  "events": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "maxLength": 255
                    },
                    "minItems": 1
                  },
                  "active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "`WebhookEndpointResource`",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/WebhookEndpointResource"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationException"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationException"
          }
        }
      },
      "delete": {
        "operationId": "v1.webhook-endpoints.destroy",
        "description": "Deliveries already recorded go with it.",
        "summary": "Delete a webhook endpoint",
        "tags": ["Webhook endpoints", "WebhookEndpoint"],
        "parameters": [
          {
            "name": "endpoint",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationException"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "http": {
        "type": "http",
        "description": "A team API key, sent as `Authorization: Bearer sk_live_\u2026`. Keys are issued per team and act only on that team.",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "WebhookEndpointResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The endpoint's stable identifier."
          },
          "url": {
            "type": "string",
            "description": "Where deliveries are POSTed."
          },
          "description": {
            "type": ["string", "null"]
          },
          "events": {
            "type": "array",
            "description": "Event types this endpoint receives. `*` means every type, including ones added later.",
            "items": {
              "type": "string"
            }
          },
          "active": {
            "type": "boolean",
            "description": "Whether deliveries are currently being sent."
          },
          "secret": {
            "type": "string",
            "description": "The signing secret. Returned only when the endpoint is created \u2014 it is not recoverable\nafterwards, and every delivery is signed with it."
          },
          "created_at": {
            "type": ["string", "null"]
          }
        },
        "required": ["id", "url", "description", "events", "active", "created_at"],
        "title": "WebhookEndpointResource"
      }
    },
    "responses": {
      "AuthenticationException": {
        "description": "Unauthenticated",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Error overview."
                }
              },
              "required": ["message"]
            }
          }
        }
      },
...
      "ValidationException": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Errors overview."
                },
                "errors": {
                  "type": "object",
                  "description": "A detailed description of each field that failed validation.",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "required": ["message", "errors"]
            }
          }
        }
      }
    }
  }
}